pyproject.toml
[tool.poetry]
name = "cmpnyinfo"
version = "0.1.0"
description = ""
authors = ["rogerkimjazzlover <minseungkim1017@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
beautifulsoup4 = "4.12.2"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
contourpy = "1.2.0"
cycler = "0.12.1"
django = "5.0.1"
djangorestframework = "3.14.0"
exceptiongroup = "1.2.0"
finance-datareader = "0.9.66"
fonttools = "4.47.2"
gevent = "24.2.1"
greenlet = "3.0.3"
gunicorn = "21.2.0"
idna = "3.6"
iniconfig = "2.0.0"
kiwisolver = "1.4.5"
lxml = "5.1.0"
matplotlib = "3.8.2"
numpy = "1.26.3"
opendartreader = "0.2.3"
packaging = "23.2"
pandas = "2.1.4"
pillow = "10.2.0"
plotly = "5.18.0"
pluggy = "1.4.0"
pyparsing = "3.1.1"
pytest = "8.0.0"
pytest-django = "4.8.0"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
requests = "2.31.0"
requests-file = "1.5.1"
six = "1.16.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
tabulate = "0.9.0"
tenacity = "8.2.3"
tomli = "2.0.1"
tqdm = "4.66.1"
typing-extensions = "4.9.0"
tzdata = "2023.4"
urllib3 = "2.1.0"
zope-event = "5.0"
zope-interface = "6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
